+++ packages/0xsequence/tests/browser/wallet-provider/dapp2.test.ts
@@ -1,10 +1,7 @@
 import { test, assert } from '../../utils/assert'
 import { ethers, Wallet as EOAWallet } from 'ethers'
 import { TypedDataDomain, TypedDataField } from '@ethersproject/abstract-signer'
-import { Wallet, DefaultProviderConfig, Web3Provider } from '@0xsequence/provider'
-import { sequenceContext, WalletContext, JsonRpcSender, JsonRpcRequest, JsonRpcResponseCallback } from '@0xsequence/network'
-import { isValidSignature, recoverConfig } from '@0xsequence/wallet'
-import { addressOf } from '@0xsequence/config'
+import { Wallet, DefaultProviderConfig } from '@0xsequence/provider'
 import { configureLogger, packMessageData } from '@0xsequence/utils'
 import { testAccounts, getEOAWallet, deployWalletContext, testWalletContext, sendETH } from '../testutils'