+++ packages/auth/tests/session.spec.ts
@@ -6,7 +6,6 @@ import { CallReceiverMock, HookCallerMock } from '@0xsequence/wallet-contracts'
 import { LocalRelayer } from '@0xsequence/relayer'
 
 import { WalletContext, NetworkConfig } from '@0xsequence/network'
-import { JsonRpcProvider } from '@ethersproject/providers'
 import { ethers, Signer as AbstractSigner } from 'ethers'
 
 import chaiAsPromised from 'chai-as-promised'
@@ -26,7 +25,7 @@ import { ETHAuth } from '@0xsequence/ethauth'
 type EthereumInstance = {
   chainId?: number
   providerUrl?: string
-  provider?: JsonRpcProvider
+  provider?: ethers.providers.JsonRpcProvider
   signer?: AbstractSigner
 }
 
@@ -72,7 +71,7 @@ describe('Wallet integration', function () {
 
   before(async () => {
     // Provider from hardhat without a server instance
-    ethnode.providerUrl = `http://localhost:9546/`
+    ethnode.providerUrl = `http://127.0.0.1:9546/`
     ethnode.provider = new ethers.providers.JsonRpcProvider(ethnode.providerUrl)
 
     ethnode.signer = ethnode.provider.getSigner()
@@ -408,7 +407,7 @@ describe('Wallet integration', function () {
       }
     })
 
-    const configFinder = new SequenceUtilsFinder(networks.find(n => n.isAuthChain).provider)
+    const configFinder = new SequenceUtilsFinder(networks.find(n => n.isAuthChain)!.provider!)
 
     const session = Session.open({
       sequenceApiUrl: '',
@@ -442,28 +441,27 @@ describe('Wallet integration', function () {
 
     let alwaysFail: boolean = false
 
-    const sequenceApiUrl = 'http://localhost:8099'
+    const sequenceApiUrl = 'http://127.0.0.1:8099'
 
     beforeEach(() => {
       fakeJwt = ethers.utils.hexlify(ethers.utils.randomBytes(64))
 
       server = mockServer.getLocal()
       server.start(8099)
-
-      server.post('/rpc/API/GetAuthToken').thenCallback(async request => {
+      server.forPost('/rpc/API/GetAuthToken').thenCallback(async request => {
         if (delayMs !== 0) await delay(delayMs)
 
         const ethauth = new ETHAuth(ValidateSequenceUndeployedWalletProof(context), ValidateSequenceDeployedWalletProof)
 
-        ethauth.chainId = ethnode.chainId
-        ethauth.configJsonRpcProvider(ethnode.providerUrl)
+        ethauth.chainId = ethnode.chainId!
+        ethauth.configJsonRpcProvider(ethnode.providerUrl!)
 
         totalCount++
 
         if (alwaysFail) return { statusCode: 400 }
 
         try {
-          const proof = await ethauth.decodeProof(request.body.json['ewtString'])
+          const proof = await ethauth.decodeProof((await request.body.getJson())!['ewtString'])
           proofAddress = ethers.utils.getAddress(proof.address)
 
           if (recoverCount[proofAddress]) {
@@ -660,8 +658,8 @@ describe('Wallet integration', function () {
 
       expect(await session._jwt?.token).to.equal(fakeJwt)
 
-      server.post('/rpc/API/FriendList').thenCallback(async request => {
-        const hasToken = request.headers['authorization'].includes(fakeJwt)
+      server.forPost('/rpc/API/FriendList').thenCallback(async request => {
+        const hasToken = request.headers['authorization']!.includes(fakeJwt)
         return { statusCode: hasToken ? 200 : 401, body: JSON.stringify({}) }
       })
 
@@ -710,8 +708,8 @@ describe('Wallet integration', function () {
 
       expect(await session._jwt?.token).to.equal(fakeJwt)
 
-      server.post('/rpc/API/FriendList').thenCallback(async request => {
-        const hasToken = request.headers['authorization'].includes(fakeJwt)
+      server.forPost('/rpc/API/FriendList').thenCallback(async request => {
+        const hasToken = request.headers['authorization']!.includes(fakeJwt)
         return { statusCode: hasToken ? 200 : 401, body: JSON.stringify({}) }
       })
 
@@ -858,8 +856,8 @@ describe('Wallet integration', function () {
 
       expect(await session._jwt?.token).to.equal(fakeJwt)
 
-      server.post('/rpc/API/FriendList').thenCallback(async request => {
-        const hasToken = request.headers['authorization'].includes(fakeJwt)
+      server.forPost('/rpc/API/FriendList').thenCallback(async request => {
+        const hasToken = request.headers['authorization']!.includes(fakeJwt)
         return { statusCode: hasToken ? 200 : 401, body: JSON.stringify({}) }
       })
 
@@ -971,7 +969,7 @@ describe('Wallet integration', function () {
       totalCount = 0
 
       // Create a bunch of API clients concurrently
-      const requests = []
+      const requests: any[] = []
       while (requests.length < 10) {
         requests.push(session.getAPIClient())
       }
@@ -1038,7 +1036,7 @@ describe('Wallet integration', function () {
       const api = await session.getAPIClient()
 
       const okResponses = [true]
-      server.post('/rpc/API/FriendList').thenCallback(async () => {
+      server.forPost('/rpc/API/FriendList').thenCallback(async () => {
         return { statusCode: okResponses.shift() ? 200 : 401, body: JSON.stringify({}) }
       })