+++ packages/deployer/hardhat.config.ts
@@ -23,8 +23,6 @@ module.exports = {
     tests: './src/tests'
   },
   networks: {
-    // rinkeby: networkConfig('rinkeby'),
-    // kovan: networkConfig('kovan'),
     goerli: networkConfig('goerli'),
     mumbai: networkConfig('mumbai'),
     matic: networkConfig('matic'),