+++ packages/estimator/tests/sequence-estimator.spec.ts
@@ -5,8 +5,7 @@ import { Transaction } from '@0xsequence/transactions'
 import { LocalRelayer } from '@0xsequence/relayer'
 
 import { WalletContext, NetworkConfig } from '@0xsequence/network'
-import { JsonRpcProvider } from '@ethersproject/providers'
-import { ethers, Signer as AbstractSigner } from 'ethers'
+import { ethers, Signer as AbstractSigner, providers } from 'ethers'
 
 import { configureLogger } from '@0xsequence/utils'
 
@@ -28,7 +27,7 @@ import { encodeData } from '@0xsequence/wallet/tests/utils'
 
 type EthereumInstance = {
   chainId: number
-  provider: JsonRpcProvider
+  provider: providers.JsonRpcProvider
   signer: AbstractSigner
 }