+++ packages/wallet/src/account.ts
@@ -1,7 +1,4 @@
-import { TransactionRequest, JsonRpcProvider, Provider } from '@ethersproject/providers'
-import { Signer as AbstractSigner, BytesLike } from 'ethers'
-import { TypedDataDomain, TypedDataField } from '@ethersproject/abstract-signer'
-import { Deferrable } from '@ethersproject/properties'
+import { Signer as AbstractSigner, BytesLike, providers, utils, TypedDataDomain, TypedDataField } from 'ethers'
 import { Signer, NotEnoughSigners, SignedTransactionsCallback } from './signer'
 import {
   SignedTransactions,
@@ -55,7 +52,7 @@ export class Account extends Signer {
 
   // provider points at the main chain for compatability with the Signer.
   // Use getProvider(chainId) to get the provider for the respective network.
-  provider: JsonRpcProvider
+  provider: providers.JsonRpcProvider
 
   // memoized value
   private _chainId?: number
@@ -163,7 +160,7 @@ export class Account extends Signer {
     return this._wallets[0].wallet.getSigners()
   }
 
-  async getProvider(chainId?: number): Promise<JsonRpcProvider | undefined> {
+  async getProvider(chainId?: number): Promise<providers.JsonRpcProvider | undefined> {
     if (!chainId) return this.mainWallet()?.wallet.getProvider()
     return this._wallets.find(w => w.network.chainId === chainId)?.wallet.getProvider()
   }
@@ -269,7 +266,7 @@ export class Account extends Signer {
   }
 
   async getFeeOptions(
-    transaction: Deferrable<Transactionish>,
+    transaction: utils.Deferrable<Transactionish>,
     chainId?: ChainIdLike,
     allSigners: boolean = true
   ): Promise<{ options: FeeOption[]; quote?: FeeQuote }> {
@@ -306,7 +303,7 @@ export class Account extends Signer {
   }
 
   async sendTransaction(
-    dtransactionish: Deferrable<Transactionish>,
+    dtransactionish: utils.Deferrable<Transactionish>,
     chainId?: ChainIdLike,
     allSigners: boolean = true,
     quote?: FeeQuote,
@@ -325,7 +322,7 @@ export class Account extends Signer {
   }
 
   async sendTransactionBatch(
-    transactions: Deferrable<TransactionRequest[] | Transaction[]>,
+    transactions: utils.Deferrable<providers.TransactionRequest[] | Transaction[]>,
     chainId?: ChainIdLike,
     allSigners: boolean = true,
     quote?: FeeQuote,
@@ -335,7 +332,7 @@ export class Account extends Signer {
   }
 
   async signTransactions(
-    dtransactionish: Deferrable<Transactionish>,
+    dtransactionish: utils.Deferrable<Transactionish>,
     chainId?: ChainIdLike,
     allSigners?: boolean
   ): Promise<SignedTransactions> {
@@ -354,7 +351,7 @@ export class Account extends Signer {
   }
 
   async prependConfigUpdate(
-    dtransactionish: Deferrable<Transactionish>,
+    dtransactionish: utils.Deferrable<Transactionish>,
     chainId?: ChainIdLike,
     allSigners?: boolean,
     skipThresholdCheck?: boolean
@@ -616,11 +613,11 @@ export class Account extends Signer {
     return this.options.networks.find(network => network.isDefaultChain)!.chainId
   }
 
-  connect(_: Provider): AbstractSigner {
+  connect(_: providers.Provider): AbstractSigner {
     throw new Error('connect method is not supported in MultiWallet')
   }
 
-  signTransaction(_: Deferrable<TransactionRequest>): Promise<string> {
+  signTransaction(_: utils.Deferrable<providers.TransactionRequest>): Promise<string> {
     throw new Error('signTransaction method is not supported in MultiWallet, please use signTransactions(...)')
   }
 }